feat: introduce MessageBuilder
