+++ lib/fromZodError.ts
@@ -1,23 +1,22 @@
-import {
-  ISSUE_SEPARATOR,
-  MAX_ISSUES_IN_MESSAGE,
-  PREFIX,
-  PREFIX_SEPARATOR,
-  UNION_SEPARATOR,
-} from './config.ts';
-import { getMessageFromZodIssue } from './fromZodIssue.ts';
-import { prefixMessage } from './prefixMessage.ts';
-import { ValidationError } from './ValidationError.ts';
 import { fromError } from './fromError.ts';
 import { isZodErrorLike } from './isZodErrorLike.ts';
+import {
+  createMessageBuilder,
+  type CreateMessageBuilderProps,
+  type MessageBuilder,
+} from './MessageBuilder.ts';
+import { isNonEmptyArray } from './utils/NonEmptyArray.ts';
+import { ValidationError } from './ValidationError.ts';
 import type * as zod from 'zod';
-import type { FromZodIssueOptions } from './fromZodIssue.ts';
 
 export type ZodError = zod.ZodError;
 
-export type FromZodErrorOptions = FromZodIssueOptions & {
-  maxIssuesInMessage?: number;
-};
+export type FromZodErrorOptions =
+  | {
+      messageBuilder: MessageBuilder;
+    }
+  // maintain backwards compatibility
+  | CreateMessageBuilderProps;
 
 export function fromZodError(
   zodError: ZodError,
@@ -38,36 +37,25 @@ export function fromZodErrorWithoutRuntimeCheck(
   zodError: ZodError,
   options: FromZodErrorOptions = {}
 ): ValidationError {
-  const {
-    maxIssuesInMessage = MAX_ISSUES_IN_MESSAGE,
-    issueSeparator = ISSUE_SEPARATOR,
-    unionSeparator = UNION_SEPARATOR,
-    prefixSeparator = PREFIX_SEPARATOR,
-    prefix = PREFIX,
-    includePath = true,
-  } = options;
-
   const zodIssues = zodError.errors;
 
-  const reason =
-    zodIssues.length === 0
-      ? zodError.message
-      : zodIssues
-          // limit max number of issues printed in the reason section
-          .slice(0, maxIssuesInMessage)
-          // format error message
-          .map((issue) =>
-            getMessageFromZodIssue({
-              issue,
-              issueSeparator,
-              unionSeparator,
-              includePath,
-            })
-          )
-          // concat as string
-          .join(issueSeparator);
-
-  const message = prefixMessage(reason, prefix, prefixSeparator);
+  let message: string;
+  if (isNonEmptyArray(zodIssues)) {
+    const messageBuilder = createMessageBuilderFromOptions(options);
+    message = messageBuilder(zodIssues);
+  } else {
+    message = zodError.message;
+  }
 
   return new ValidationError(message, { cause: zodError });
+}
+
+function createMessageBuilderFromOptions(
+  options: FromZodErrorOptions
+): MessageBuilder {
+  if ('messageBuilder' in options) {
+    return options.messageBuilder;
+  }
+
+  return createMessageBuilder(options);
 }
