+++ lib/fromZodIssue.test.ts
@@ -5,7 +5,7 @@ import { ValidationError } from './ValidationError.ts';
 import { isZodErrorLike } from './isZodErrorLike.ts';
 
 describe('fromZodIssue()', () => {
-  test('handles zod.string() schema errors', () => {
+  test('handles ZodIssue', () => {
     const schema = zod.string().email();
 
     try {
@@ -27,104 +27,6 @@ describe('fromZodIssue()', () => {
               },
             ]
         `);
-      }
-    }
-  });
-
-  test('handles zod.function() argument errors', () => {
-    const fn = zod
-      .function()
-      .args(zod.number())
-      .implement((num) => num * 2);
-
-    try {
-      // @ts-expect-error Intentionally wrong to exercise runtime checking
-      fn('foo');
-    } catch (err) {
-      if (isZodErrorLike(err)) {
-        const validationError = fromZodIssue(err.issues[0]);
-        expect(validationError).toBeInstanceOf(ValidationError);
-        expect(validationError.message).toMatchInlineSnapshot(
-          `"Validation error: Invalid function arguments; Expected number, received string at index 0"`
-        );
-        expect(validationError.details).toMatchInlineSnapshot(`
-          [
-            {
-              "argumentsError": [ZodError: [
-            {
-              "code": "invalid_type",
-              "expected": "number",
-              "received": "string",
-              "path": [
-                0
-              ],
-              "message": "Expected number, received string"
-            }
-          ]],
-              "code": "invalid_arguments",
-              "message": "Invalid function arguments",
-              "path": [],
-            },
-          ]
-        `);
-      }
-    }
-  });
-
-  test('handles zod.function() return value errors', () => {
-    const fn = zod
-      .function()
-      .returns(zod.number())
-      // @ts-expect-error Intentionally wrong to exercise runtime checking
-      .implement(() => 'foo');
-
-    try {
-      fn();
-    } catch (err) {
-      if (isZodErrorLike(err)) {
-        const validationError = fromZodIssue(err.issues[0]);
-        expect(validationError).toBeInstanceOf(ValidationError);
-        expect(validationError.message).toMatchInlineSnapshot(
-          `"Validation error: Invalid function return type; Expected number, received string"`
-        );
-        expect(validationError.details).toMatchInlineSnapshot(`
-          [
-            {
-              "code": "invalid_return_type",
-              "message": "Invalid function return type",
-              "path": [],
-              "returnTypeError": [ZodError: [
-            {
-              "code": "invalid_type",
-              "expected": "number",
-              "received": "string",
-              "path": [],
-              "message": "Expected number, received string"
-            }
-          ]],
-            },
-          ]
-        `);
-      }
-    }
-  });
-
-  test('respects `includePath` prop when set to `false`', () => {
-    const schema = zod.object({
-      name: zod.string().min(3, '"Name" must be at least 3 characters'),
-    });
-
-    try {
-      schema.parse({ name: 'jo' });
-    } catch (err) {
-      if (isZodErrorLike(err)) {
-        const validationError = fromZodIssue(err.issues[0], {
-          includePath: false,
-        });
-        expect(validationError).toBeInstanceOf(ValidationError);
-        expect(validationError.message).toMatchInlineSnapshot(
-          `"Validation error: "Name" must be at least 3 characters"`
-        );
       }
     }
   });
